[ovsp4rt] Support 24-bit VNI values
- Widened tunnel_info.vni from uint16_t to uint32_t.

- Modified EncodeVniValue() to encode the low-order 24 bits of a
  uint32_t value.

- Implemented a new EncodeTunnelId() function to use in place of
  EncodeVniValue() when the destination is a tunnel_id field. It
  encodes the low-order 20 bits of a uint32_t value.

- Modified unit tests to use uint32_t instead of uint16_t to
  represent a VNI value. Changed the ipv4 VNI test value from
  16 to 24 bits.

+++ ovs-p4rt/sidecar/testing/ipv6_tunnel_test.h
@@ -18,7 +18,7 @@ constexpr int IPV6_PREFIX_LEN = 64;
 
 constexpr uint16_t SRC_PORT = 0x1984;
 constexpr uint16_t DST_PORT = 0x4224;
-constexpr uint16_t VNI = 0x1066;
+constexpr uint32_t VNI = 0x1066;
 
 class Ipv6TunnelTest : public TableEntryTest {
  protected:
