+++ ovs-p4rt/sidecar/testing/vxlan_encap_v4_vlan_pop_test.cc
@@ -50,7 +50,7 @@ TEST_F(Ipv4TunnelTest, vxlan_encap_v4_vlan_pop_params_are_correct) {
 
   absl::optional<uint16_t> src_port;
   absl::optional<uint16_t> dst_port;
-  absl::optional<uint16_t> vni;
+  absl::optional<uint32_t> vni;
 
   for (int i = 0; i < num_params; ++i) {
     auto param = params[i];
